+++ src/app/dashboard/map/map.component.ts
@@ -3,7 +3,6 @@ import { MapService } from './map.service'
 
 @Component({
     selector: 'app-map',
-    imports: [],
     templateUrl: './map.component.html',
     styleUrls: ['./map.component.scss'],
     standalone: true
@@ -12,6 +11,6 @@ export class MapComponent implements AfterViewInit {
     constructor(private mapService: MapService) {}
 
     ngAfterViewInit(): void {
-        this.mapService.initMap()
+        this.mapService.initMap('main-map')
     }
 }