+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 - Now supporting mapping/grouping input fields ([#145](https://gitlab.heigit.org/climate-action/web-app/-/issues/145))
 
+### Added
+
+- Added a report builder to the dashboard to view multiple artifacts at once ([#210](https://gitlab.heigit.org/climate-action/web-app/-/issues/210))
+
 ## [2.1.0](https://gitlab.heigit.org/climate-action/web-app/-/releases/2.1.0) - 2025-01-29
 
 ### Changed